+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Bug where Pill Schedules "Last 10 Days" and "Last 20 Days" would use the wrong start date of the month.
+
 ### Changed
 
 - iPad font size for Hormone Cell text was decreased.